Fix(?) GH check fail. Some type shuffling/consolidation.
CI failure: https://github.com/elastic/kibana/pull/37950/checks?check_run_id=141144417
