+++ x-pack/plugins/integrations_manager/server/index.ts
@@ -3,14 +3,12 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-
-import { PluginInitializerContext } from 'src/core/server';
-import { CoreSetup } from 'src/core/server';
+import { CoreSetup } from '../common/types';
 import { routes } from './routes';
 
 export class Plugin {
-  public setup(core: CoreSetup, initializerContext: PluginInitializerContext) {
-    const { server } = core.http;
-    routes.forEach(route => server.route(route));
+  public setup(core: CoreSetup) {
+    const { route } = core.http;
+    routes.forEach(definition => route(definition));
   }
 }
